+++ Kernel/cpan-lib/Locale/Codes/LangExt_Codes.pm
@@ -1,36 +1,9 @@
-package Locale::Codes::LangExt_Codes;
+package #
+Locale::Codes::LangExt_Codes;
 
 # This file was automatically generated.  Any changes to this file will
 # be lost the next time 'get_codes' is run.
-#    Generated on: Mon Jun  3 10:41:30 EDT 2013
-
-=pod
-
-=head1 NAME
-
-Locale::Codes::LangExt_Codes - langext codes for the Locale::Codes::LangExt module
-
-=head1 SYNOPSIS
-
-This module contains data used by the Locale::Codes::LangExt module. It is
-not intended to be used directly, and contains no callable routines.
-
-=head1 AUTHOR
-
-See Locale::Codes for full author history.
-
-Currently maintained by Sullivan Beck (sbeck@cpan.org).
-
-=head1 COPYRIGHT
-
-   Copyright (c) 1997-2001 Canon Research Centre Europe (CRE).
-   Copyright (c) 2001-2010 Neil Bowers
-   Copyright (c) 2010-2013 Sullivan Beck
-
-This module is free software; you can redistribute it and/or
-modify it under the same terms as Perl itself.
-
-=cut
+#    Generated on: Thu Jan 23 14:29:32 EST 2014
 
 use strict;
 require 5.006;
@@ -38,7 +11,7 @@ use warnings;
 use utf8;
 
 our($VERSION);
-$VERSION='3.26';
+$VERSION='3.29';
 
 $Locale::Codes::Data{'langext'}{'id'} = '0228';